requirements.txt
streamlit
faster-whisper==1.0.3
ffmpeg-python
numpy
